WIP: Update examples to use traces. NOTE: 3 build errors still
