+++ examples/alarmBot/abilities.ts
@@ -14,51 +14,19 @@ export interface UserState {
 export const abilities = [
   {
     name: 'greeting',
-    slots: [],
+    traces: [],
     onComplete: () => {
       return 'Hello! Welcome to alarm bot.'
     }
   },
   {
     name: 'addAlarm',
-    slots: [
+    traces: [
       {
-        name: 'alarmName',
-        query: () => { return 'What is the name of the alarm?' },
-        retry: (submittedValue, convoStorageLayer, turnCount) => {
-          const phrase = [`Please try a new name (attempt: ${turnCount})`, `Try harder.. (attempt: ${turnCount})`]
-          if (turnCount > phrase.length - 1) {
-            return phrase[phrase.length - 1]
-          }
-          return phrase[turnCount]
-        },
-        validate: (value) => {
-          if (value.toLowerCase() === 'hao') {
-            return { isValid: false, reason: `${value} is not a good alarm name.` }
-          }
-          return { isValid: true, reason: null }
-        },
-        onFill: (value) => `ok! name is set to ${value}.`
+        slotName: 'alarmName'
       },
       {
-        name: 'alarmTime',
-        query: () => { return 'What is the time you want to set?' },
-        retry: (submittedValue, convoStorageLayer, turnCount) => {
-          const phrases: string[] = ['let\'s try again', 'what is the time you want to set?']
-          return randomElement(phrases)
-        },
-        validate: (value: string) => {
-          if (!value.toLowerCase().endsWith('pm') && !value.toLowerCase().endsWith('am')) {
-            return {
-              isValid: false,
-              reason: 'Needs to set PM or AM',
-            }
-          }
-          return {
-            isValid: true
-          }
-        },
-        onFill: (value) => `ok! time is set to ${value}.`
+        slotName: 'alarmTime'
       }
     ],
     onComplete: async (submittedData, {read, save}) => {
@@ -79,12 +47,9 @@ export const abilities = [
   },
   {
     name: 'removeAlarm',
-    slots: [
+    traces: [
       {
-        name: 'alarmName',
-        query: () => {
-          return 'What is the name of the alarm you would like to remove?'
-        }
+        slotName: 'alarmName'
       }
     ],
     onComplete: async (submittedData, {read, save}) => {
@@ -110,7 +75,7 @@ export const abilities = [
   },
   {
     name: 'listAlarms',
-    slots: [],
+    traces: [],
     onComplete: async (submittedData, {read}) => {
       const convoState = await read()
       const alarms = convoState.alarms || []
@@ -123,7 +88,7 @@ export const abilities = [
   },
   {
     name: 'listAbility',
-    slots: [],
+    traces: [],
     onComplete: (submittedData, convoStorageLayer, { getAbilityList }) => {
       const abilityList = getAbilityList()
       const abilities = abilityList.map((ability) => ability.name).join(', ')
