+++ examples/echoBot/bot.ts
@@ -2,13 +2,19 @@
 // Licensed under the MIT License.
 
 import { TurnContext } from 'botbuilder'
+import { StorageLayerType } from '../../src'
 import * as wolf from 'wolf-core'
-import {} from '../../src'
-import abilities from './abilities'
+import { abilities } from './abilities'
+import { slots } from './slots'
 import nlp from './nlp'
 
 export interface ConversationData {
   name?: string
+}
+
+const flow: wolf.Flow<ConversationData, StorageLayerType<ConversationData>> = {
+  abilities,
+  slots
 }
 
 export class MyBot {
@@ -37,7 +43,7 @@ export class MyBot {
       this.wolfStorageLayer(turnContext),
       this.conversationStorageLayer(turnContext, {}),
       () => nlp(turnContext),
-      () => abilities,
+      () => flow,
       'echo'
     )
